+++ src/client/routes/admin/NotificationSettingsSection.tsx
@@ -1,0 +1,60 @@
+import { toast } from 'sonner';
+import { trpc } from '@/client/lib/trpc';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Label } from '@/components/ui/label';
+import { Skeleton } from '@/components/ui/skeleton';
+import { Switch } from '@/components/ui/switch';
+
+export function NotificationSettingsSection() {
+  const { data: settings, isLoading } = trpc.userSettings.get.useQuery();
+  const utils = trpc.useUtils();
+  const updateSettings = trpc.userSettings.update.useMutation({
+    onSuccess: () => {
+      utils.userSettings.get.invalidate();
+    },
+    onError: (error) => {
+      toast.error(`Failed to update settings: ${error.message}`);
+    },
+  });
+
+  const handleToggleSound = (checked: boolean) => {
+    updateSettings.mutate({ playSoundOnComplete: checked });
+  };
+
+  if (isLoading) {
+    return (
+      <Card>
+        <CardHeader>
+          <CardTitle>Notification Settings</CardTitle>
+          <CardDescription>Configure notification behavior</CardDescription>
+        </CardHeader>
+        <CardContent>
+          <Skeleton className="h-10 w-full" />
+        </CardContent>
+      </Card>
+    );
+  }
+
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle>Notification Settings</CardTitle>
+        <CardDescription>Configure notification behavior</CardDescription>
+      </CardHeader>
+      <CardContent className="space-y-4">
+        <div className="flex flex-col gap-2 sm:flex-row sm:items-center sm:justify-between">
+          <div className="space-y-0.5">
+            <Label htmlFor="sound-toggle">Play completion sound</Label>
+            <p className="text-sm text-muted-foreground">Play a sound when a workspace finishes</p>
+          </div>
+          <Switch
+            id="sound-toggle"
+            checked={settings?.playSoundOnComplete ?? true}
+            onCheckedChange={handleToggleSound}
+            disabled={updateSettings.isPending}
+          />
+        </div>
+      </CardContent>
+    </Card>
+  );
+}
