+++ src/client/routes/admin/index.ts
@@ -1,5 +1,11 @@
 export { type ApiUsageData, ApiUsageSection, type ApiUsageSectionProps } from './ApiUsageSection';
+export { AppInfoSection } from './AppInfoSection';
+export { ChatProviderDefaultsSection } from './ChatProviderDefaultsSection';
+export { CliAuthSection } from './CliAuthSection';
+export { DataBackupSection } from './DataBackupSection';
+export { IdeSettingsSection } from './IdeSettingsSection';
 export { IssueTrackingSection, ProjectIssueTrackingCard } from './IssueTrackingSection';
+export { NotificationSettingsSection } from './NotificationSettingsSection';
 export { PeriodicTasksSection } from './PeriodicTasksSection';
 export {
   type ProcessesData,
@@ -7,5 +13,8 @@ export {
   type ProcessesSectionProps,
   ProcessesSectionSkeleton,
 } from './ProcessesSection';
+export { ProjectSettingsSection } from './ProjectSettingsSection';
+export { RatchetSettingsSection } from './RatchetSettingsSection';
+export { ServerLogsSection } from './ServerLogsSection';
 export { StatCard, type StatCardProps } from './StatCard';
 export { VoiceModeSection } from './VoiceModeSection';
